Moving UseOfForce
